requirements.txt
paho-mqtt==1.5.1
playsound==1.3.0
pulsectl==21.9.1
PyAudio==0.2.11
